rename autoscaling/metrics to autoscaling/datasource
